+++ R/transmute_nse.R
@@ -8,6 +8,7 @@
 #' @param .data data.frame
 #' @param ... stringified expressions to transmute by.
 #' @param transmute_nse_env environment to work in.
+#' @param transmute_nse_warn logical, if TRUE warn about possible name collisions.
 #' @return .data with altered columns(other columns dropped).
 #'
 #' @examples
@@ -19,13 +20,17 @@
 #'
 #' @export
 #'
-transmute_nse <- function(.data, ...,  transmute_nse_env = parent.frame()) {
+transmute_nse <- function(.data, ...,
+                          transmute_nse_env = parent.frame(),
+                          transmute_nse_warn = TRUE) {
   transmuteTerms <- wrapr::qae(...)
   force(transmute_nse_env)
   res <- .data
   len <- length(transmuteTerms)
   if(len>1) {
-    res <- transmute_se(res, transmuteTerms, env=transmute_nse_env)
+    res <- transmute_se(res, transmuteTerms,
+                        env = transmute_nse_env,
+                        warn = transmute_nse_warn)
   }
   res
 }